housekeeping: moved script location to folder `general`
