+++ scripts/general/gapminder_data_analysis.py
